+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add Pyramid type.
 - Clean up assay names.
 - Add new assay_type and remove portal_uploaded_protocol_files.
+- Use YAML rather than TSV as the base format.
 
 ## v0.0.2 - 2020-06-29
 - Add lightsheet.
