+++ README.md
@@ -26,8 +26,8 @@ pip install -r requirements-dev.txt
 ./test.sh
 ```
 
-- Schemas begin as TSV files.
-- The information is then consolidated into a single YAML by `definitions-tsv-to-yaml.py`.
+- Schemas begin as YAML files.
+- The information is then consolidated into a single YAML by `consolidate-yaml.py`.
 - And from that single YAML, JSON Schemas are generated by `definitions-yaml-to-schema.py`.
 
 ### Update
