--- data/definitions/enums/assay_types.tsv
@@ -1,27 +1,0 @@
-Value	Description
-AF	Autofluorescence Microscopy
-ATACseq-bulk	Bulk ATAC-seq
-MxIF	Multiplexed IF Microscopy
-CODEX	CODEX
-IMC	Imaging Mass Cytometry
-MALDI-IMS-neg	MALDI IMS negative
-MALDI-IMS-pos	MALDI IMS positive
-PAS	PAS Stained Microscopy
-bulk-RNA	Bulk RNA-seq
-SNAREseq	SNARE-seq
-TMT-LC-MS	TMT LC-MS
-Targeted-Shotgun-LC-MS	Targeted Shotgun / Flow-injection LC-MS
-LC-MS-untargeted	Untargeted LC-MS
-WGS	Whole Genome Sequencing
-scRNA-Seq-10x	scRNA-seq (10x Genomics)
-sciATACseq	sciATAC-seq
-sciRNAseq	sciRNA-seq
-seqFish	seqFish
-seqFish_lab_processed	processed seqFISH data [processed at source lab]
-snATACseq	snATAC-seq
-snRNAseq	snRNA-seq
-codex_cytokit	CODEX [Cytokit + SPRM]
-salmon_rnaseq_10x	scRNA-seq (10x Genomics) [Salmon]
-Lightsheet	Lightsheet Microscopy
-image_pyramid	Image Pyramid
-
