--- data/definitions/enums/dataset_status_types.tsv
@@ -1,12 +1,0 @@
-Value	Description
-NEW	new
-REOPENED	reopened
-INVALID	invalid
-QA	qa
-LOCKED	locked
-PROCESSING	processing
-PUBLISHED	published
-UNPUBLISHED	unpublished
-DEPRECATED	deprecated
-ERROR	error
-HOLD	hold
