--- data/definitions/enums/entity_types.tsv
@@ -1,4 +1,0 @@
-Value	Description
-Dataset	A collection of data
-Donor	A donor who tissue was procured from.
-Sample	A tissue sample, could be an organ
