+++ data/definitions/enums/entity_types.yaml
@@ -1,0 +1,6 @@
+Dataset:
+    description: A collection of data
+Donor:
+    description: A donor who tissue was procured from.
+Sample:
+    description: A tissue sample, could be an organ
