+++ data/definitions/enums/tissue_sample_types.yaml
@@ -1,0 +1,90 @@
+atacseq:
+  description: ATACseq
+biopsy:
+  description: Biopsy
+blood:
+  description: Blood
+cell_lysate:
+  description: Cell lysate
+clarity_hydrogel:
+  description: CLARITY hydrogel
+codex:
+  description: CODEX
+cryosections_curls_from_fresh_frozen_oct:
+  description: Cryosections/curls from fresh frozen OCT
+cryosections_curls_rnalater:
+  description: Cryosectinos/curls RNAlater
+ffpe_block:
+  description: FFPE block
+ffpe_slide:
+  description: FFPE slide
+fixed_frozen_section_slide:
+  description: Fixed Frozen section slide
+fixed_tissue_piece:
+  description: Fixed tissue piece
+flash_frozen_liquid_nitrogen:
+  description: Flash frozen, liquid nitrogen
+formalin_fixed_oct_block:
+  description: Formalin fixed OCT block
+fresh_frozen_oct_block:
+  description: Fresh frozen oct block
+fresh_frozen_section_slide:
+  description: Fresh Frozen section slide
+fresh_frozen_tissue:
+  description: Fresh frozen tissue
+fresh_frozen_tissue_section:
+  description: Fresh Frozen Tissue Section
+fresh_tissue:
+  description: Fresh tissue
+frozen_cell_pellet_buffy_coat:
+  description: Frozen cell pellet (Buffy coat)
+gdna:
+  description: gDNA
+module:
+  description: Module
+nuclei:
+  description: Nuclei
+nuclei_rnalater:
+  description: Nuclei RNAlater
+organ:
+  description: Organ
+organ_piece:
+  description: Organ Piece
+other:
+  description: Other
+pbmc:
+  description: PBMC
+pfa_fixed_frozen_oct_block:
+  description: PFA Fixed frozen OCT block
+plasma:
+  description: Plasma
+protein:
+  description: Protein
+ran_poly_a_enriched:
+  description: RNA, poly-A enriched
+rna_total:
+  description: RNA, total
+rnalater_treated_and_stored:
+  description: RNAlater treated and stored
+rnaseq:
+  description: RNAseq
+scatacseq:
+  description: scATACseq
+scrnaseq:
+  description: scRNAseq
+segment:
+  description: Segment
+seqfish:
+  description: seqFISH
+serum:
+  description: Serum
+single_cell_cryopreserved:
+  description: Single cell cryopreserved
+snatacseq:
+  description: snATACseq
+snrnaseq:
+  description: snRNAseq
+tissue_lysate:
+  description: Tissue lysate
+wgs:
+  description: Whole Genome Sequencing
